Refactor : ReviewList 컴포넌트 이름 변경 및 위치 이동
